Rename file to connector
